fix: scipy.signal.resample_poly renvoyait du silence total sur un tableau int16
Bug réel découvert en construisant le corpus d'entraînement du futur
détecteur "Jaris" (voir commits précédents, en cours) : le corpus
ré-échantillonné (44,1 kHz -> 16 kHz) via resample_poly donnait un
modèle bloqué à 0% de rappel quel que soit l'hyperparamètre ajusté.
En creusant (comparaison du melspectrogramme brut ONNX, puis du
signal audio lui-même), les fichiers ré-échantillonnés se sont
révélés totalement SILENCIEUX (RMS = 0), pas juste mal entraînés.

Reproduit avec un script minimal, hors de tout contexte Jaris :
scipy.signal.resample_poly sur un tableau int16 renvoie 0 partout,
SANS la moindre erreur ni avertissement, quel que soit le ratio
(testé 44100->16000 ET 48000->16000) ou le contenu (testé sur un ton
pur ET du bruit aléatoire). La MÊME opération sur le MÊME signal
converti en float64 avant l'appel donne le résultat attendu.

make_audio_callback (voice_server.py) fait exactement le même appel,
dans la retombée utilisée quand un micro n'accepte pas 16 kHz
directement (micros USB/Bluetooth notamment, cas déjà documenté) :
CORRIGÉ en castant en float64 avant resample_poly. Ce chemin n'avait
jamais été vérifié en usage réel (aucun micro de ce type rencontré
par Léo jusqu'ici) -- sans ce correctif, toute la chaîne audio (mot
d'activation ET transcription) serait restée silencieuse en
permanence sur un tel micro, sans le moindre message d'erreur pour
comprendre pourquoi.

+++ python/voice_server.py
@@ -225,7 +225,14 @@ def main() -> None:
         def on_audio(indata: np.ndarray, _frames: int, _time_info, status) -> None:
             if status:
                 emit({"event": "log", "message": str(status)})
-            resampled = scipy.signal.resample_poly(indata[:, 0], up, down)
+            # scipy.signal.resample_poly sur un tableau int16 renvoie SILENCIEUSEMENT du silence total
+            # (aucune erreur, aucun avertissement) : sa filtration polyphase interne suppose une entrée en
+            # virgule flottante, un tableau entier y donne un résultat nul quel que soit le ratio ou le
+            # contenu (vérifié : ton pur ET bruit aléatoire donnent 0 en int16, un résultat correct en
+            # float64) — jamais détecté avant faute d'accès à un micro qui déclenche vraiment cette
+            # retombée (débit natif ≠ 16 kHz) en usage réel. Convertir en float AVANT le ré-échantillonnage
+            # corrige ça : c'est déjà le format attendu par les modèles ONNX de melspectrogramme en aval.
+            resampled = scipy.signal.resample_poly(indata[:, 0].astype(np.float64), up, down)
             audio_queue.put(np.clip(resampled, -32768, 32767).astype(np.int16))
 
         return on_audio
